added category service tests

+++ src/services/category.js
@@ -9,11 +9,23 @@ class CategoryService {
     return CategoryModel.find(filters);
   }
 
-  static update(filters = {}, data = {}) {
-    return CategoryModel.updateMany(filters, { $set: data });
+  static update(filters, data = {}) {
+    if (!filters) {
+      throw new Error('Filters not found');
+    }
+
+    return CategoryModel.updateMany(
+      filters,
+      { $set: data },
+      { runValidators: true }
+    );
   }
 
-  static delete(filters = {}) {
+  static delete(filters) {
+    if (!filters) {
+      throw new Error('Filters not found');
+    }
+
     return CategoryModel.deleteMany(filters);
   }
 }
